Add faceting using keywords
For this to work, we've identified "keywords" as an attribute for
faceting.

UI wise, we've added a grid layout to control the search panel. There's
a new grid area on the left where we've put our refinements list..

Currently the only facet is one keywords.

Besides the keyword facet, we also have a widget that lists the current
refinements and provides an easy way to reset those refinements.