Install eslint and prettier
This configuration uses https://github.com/wesbos/eslint-config-wesbos,
which is a solid set of defaults that runs prettier to format JS code in
conjunction with ESLint for linting.

The .prettierrc.yaml file, in addition to the config in package.json, is
needed so that prettier can be used to format other file types besides
JavaScript. Fortunately our Prettier config is simple and stable, so we
can afford the duplication.

Note I found I needed to add the eslintignore and prettierignores
because npm run lint was hanging --- apparently it was linting all of
node_modules. I haven't seen this behavior before.