Add SEO component
This uses react-helmet to inset site metadata into a page. I've also
included logo assets from the original site and
https://github.com/astropy/astropy-logo.

For the canonical URL, we might want to look into
https://www.gatsbyjs.com/plugins/gatsby-plugin-react-helmet-canonical-urls/
for an easier-to-use solution, though I don't think it emits an open
graph tag?

More work is needed to refine the social media previews.