Start styling the tutorial "hits" / cards
This is the core infrastructure for styling the Algolia hit components
for tutorials (the search result cards).

In the hits.js module we're extending the built-in Algolia hit component
so that we can pass props to the hit components in the future (e.g. to
control information display).

In the tutorialHit.js module we're setting up the result card's display
specifically for tutorial results. Right now we're just showing the
title and a snippet extract from the search.