Add ability to search for refinements
If there are too many facets in a keyword list, the user can either hit
the show more button to show additional facets, or search for the facet.

The default is to order facets by their popularity, which especially
makes sense in the context of showing results from a search term---you
want to use these keywords to filter further and it makes sense to be
able to zero in on the most popular keyword (or exclude it).

This work is using the built-in instantsearch styling so far.