Switch to using eslint and prettier separately
Previously we used eslint as a driver for prettier, as recommended by
https://github.com/wesbos/eslint-config-wesbos. Since then, prettier now
recommends running them separately
(https://prettier.io/docs/en/integrating-with-linters.html) and using
the eslint-config-prettier to turn off eslint rules that are dealt with
automatically by prettier. This means we no longer use
eslint-plugin-prettier. This also means we need to stop using
eslint-config-wesbos because that config sets up eslint-plugin-prettier.

However, we still want the rules established by eslint-config-wesbos, so
here we're carrying those rule sets over for now. It does give us the
flexibility to adjust these rules as needed, though, which is good.