Drop header with category cards and search
With the move to put the search UI on the homepage and combine results
from all content categories, we no longer want the large header with
hero search and the content category cards on the homepage. Instead, we
want that header to shrink into something more like a thin global nav
header that contains the site logo and links.