Start building up design token usage
Add more design tokens for defining a range of sizes and neutral colors,
and map those tokens to specific use-cases.

Overall, this helps maintains consistency in the design so that sizes
and colours are more consistent across the site.

In the future we'll want to build out a full grid of neutral and primary
colors to use in different situations.