Solve styled-components SSR issue
It turns out that styled-components dropped support for the @viewport
selector that was still in our boilerplate (now using a meta tag for
this anyways). But that was preventing many global styles on the :root
and html element from being parsed in the static site build.

Also moved the <GlobalStyles> outside the <StyledLayout> tag in the
Layout component because I think this is more correct (I haven't tested
whether this also contributed to the problem, in isolation, though).