Add PrioritySort virtual component
This component implements the feature where "priority" resources are
shown preferentially when a user doesn't enter a search term; but when a
search term is entered the sort automatically switches to "relevance" to
provide the best search match.

Behind the scenes, the PrioritySort component is switching between two
different Algolia indices. The relevance-based index is the default; the
priority index is a "replica" index that has a different sort order to
leverage the priority field in our Algolia records.