Add deployment/installtutorials.py script
This script downloads and installs the built tutorials artifact from
astropy/astropy-tutorials into a subdirectory of the Gatsby site build.
It can download the artifact either from a specific workflow run (as
when triggered by a tutorials release) or it can find the latest
artifact corresponding to the main branch of astropy-tutorials.